+++ packages/fighting-design/pagination/src/props.ts
@@ -16,7 +16,7 @@ export const Props = {
     require: true
   },
   /**
-   * 总数
+   * 总页数
    */
   total: {
     type: Number,
@@ -31,21 +31,21 @@ export const Props = {
     default: (): number => 10
   },
   /**
-   * 每页多少条
+   * 是否启用选择器选择最大页数
    */
   pageSizes: {
     type: Array as PropType<number[]>,
     default: (): null => null
   },
   /**
-   * 是否需要背景色
+   * 是否带有背景色
    */
   background: {
     type: Boolean,
     default: (): boolean => false
   },
   /**
-   * 是否需要圆角
+   * 是否带有圆角
    */
   round: {
     type: Boolean,