+++ packages/fighting-design/time-picker/src/props.ts
@@ -1,5 +1,14 @@
 import type { ExtractPropTypes } from 'vue'
 
-export const Props = {} as const
+export const Props = {
+  /**
+   * 绑定值
+   */
+  time: {
+    type: String,
+    default: (): string => '',
+    required: true
+  }
+} as const
 
 export type TimePickerPropsType = ExtractPropTypes<typeof Props>