+++ packages/fighting-design/toolbar/src/toolbar.vue
@@ -1,6 +1,7 @@
 <script lang="ts" setup name="FToolbar">
   import { Props } from './props'
   import { computed } from 'vue'
+  import { runCallback } from '../../_utils'
   import type { ComputedRef, CSSProperties } from 'vue'
   import type {
     ClassListInterface,
@@ -44,17 +45,21 @@
 
   /**
    * 点击触发
+   *
+   * @param evt 事件对象
    */
   const handleClick: HandleMouseEventInterface = (evt: MouseEvent): void => {
+    // 获取内部的元素节点列表
     const path: HTMLElement[] = evt.composedPath() as HTMLElement[]
 
+    // 过滤出自己的亲孩子组件
     const node: HTMLElement | undefined = path.find(
       (item: HTMLElement): boolean => item.className === 'f-toolbar-item'
     )
 
     const key: string | undefined = node ? node.dataset.key : ''
 
-    prop.click && prop.click({ evt, key } as ToolbarClickEmitInterface)
+    runCallback(prop.click, { evt, key } as ToolbarClickEmitInterface)
   }
 </script>